toolkit/components/passwordmgr/test/test_basic_form_3pw_1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Password Manager</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Password Manager test: forms with 3 password fields (form filling)
<p id="display"></p>

<div id="content" style="display: none">
  <p>The next three forms are <b>user/pass/passB/passC</b>, as all-empty, preuser(only), and preuser/pass</p>
  <form action="formtest.js" method="get">
    <input  type="text"       name="uname1"  value="">
    <input  type="password"   name="pword1"  value="">
    <input  type="password"   name="pword1B" value="">
    <input  type="password"   name="pword1C" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname2"  value="testuser2">
    <input  type="password"   name="pword2"  value="">
    <input  type="password"   name="pword2B" value="">
    <input  type="password"   name="pword2C" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname3"  value="testuser3">
    <input  type="password"   name="pword3"  value="testpass3">
    <input  type="password"   name="pword3B" value="">
    <input  type="password"   name="pword3C" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>


  <p>The next three forms are <b>user/passB/pass/passC</b>, as all-empty, preuser(only), and preuser/pass</p>
  <form action="formtest.js" method="get">
    <input  type="text"       name="uname4"  value="">
    <input  type="password"   name="pword4B" value="">
    <input  type="password"   name="pword4"  value="">
    <input  type="password"   name="pword4C" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname5"  value="testuser5">
    <input  type="password"   name="pword5B" value="">
    <input  type="password"   name="pword5"  value="">
    <input  type="password"   name="pword5C" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname6"  value="testuser6">
    <input  type="password"   name="pword6B" value="">
    <input  type="password"   name="pword6"  value="testpass6">
    <input  type="password"   name="pword6C" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <p>The next three forms are <b>user/passB/passC/pass</b>, as all-empty, preuser(only), and preuser/pass</p>
  <form action="formtest.js" method="get">
    <input  type="text"       name="uname7"  value="">
    <input  type="password"   name="pword7B" value="">
    <input  type="password"   name="pword7C" value="">
    <input  type="password"   name="pword7"  value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname8"  value="testuser8">
    <input  type="password"   name="pword8B" value="">
    <input  type="password"   name="pword8C" value="">
    <input  type="password"   name="pword8"  value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname9"  value="testuser9">
    <input  type="password"   name="pword9B" value="">
    <input  type="password"   name="pword9C" value="">
    <input  type="password"   name="pword9"  value="testpass9">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Password Manger: form fill, 3 password fields **/

// Test to make sure 3-password forms are filled properly.

function startTest() {
  // Check form 1
  is($_("uname1").value,  "testuser1", "Checking username1");
  is($_("pword1").value,  "testpass1", "Checking password1");
  is($_("pword1B").value, "",    "Checking password1B");
  is($_("pword1C").value, "",    "Checking password1C");
  // Check form 2
  is($_("uname2").value,  "testuser2", "Checking username2");
  is($_("pword2").value,  "testpass2", "Checking password2");
  is($_("pword2B").value, "",    "Checking password2B");
  is($_("pword2C").value, "",    "Checking password2C");
  // Check form 3
  is($_("uname3").value,  "testuser3", "Checking username3");
  is($_("pword3").value,  "testpass3", "Checking password3");
  is($_("pword3B").value, "",    "Checking password3B");
  is($_("pword3C").value, "",    "Checking password3C");

  // Check form 4
  is($_("uname4").value,  "testuser4", "Checking username4");
  is($_("pword4").value,  "testpass4", "Checking password4");
  is($_("pword4B").value, "",    "Checking password4B");
  is($_("pword4C").value, "",    "Checking password4C");
  // Check form 5
  is($_("uname5").value,  "testuser5", "Checking username5");
  is($_("pword5").value,  "testpass5", "Checking password5");
  is($_("pword5B").value, "",    "Checking password5B");
  is($_("pword5C").value, "",    "Checking password5C");
  // Check form 6
  is($_("uname6").value,  "testuser6", "Checking username6");
  is($_("pword6").value,  "testpass6", "Checking password6");
  is($_("pword6B").value, "",    "Checking password6B");
  is($_("pword6C").value, "",    "Checking password6C");

  // Check form 7
  is($_("uname7").value,  "testuser7", "Checking username7");
  is($_("pword7").value,  "testpass7", "Checking password7");
  is($_("pword7B").value, "",    "Checking password7B");
  is($_("pword7C").value, "",    "Checking password7C");
  // Check form 8
  is($_("uname8").value,  "testuser8", "Checking username8");
  is($_("pword8").value,  "testpass8", "Checking password8");
  is($_("pword8B").value, "",    "Checking password8B");
  is($_("pword8C").value, "",    "Checking password8C");
  // Check form 9
  is($_("uname9").value,  "testuser9", "Checking username9");
  is($_("pword9").value,  "testpass9", "Checking password9");
  is($_("pword9B").value, "",    "Checking password9B");
  is($_("pword9C").value, "",    "Checking password9C");

  // TODO: as with the 2-password cases, add tests to check for creating new
  // logins and changing passwords.
  SimpleTest.finish();
}

window.onload = startTest;

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
